Refactor subnet usage
Base subnet is now set at the top and the rest is calculated based on
this.

TODO: README need to be updated accordingly.

it's now possible to handle "n" interfaces. It is also possible to link
multiple interfaces inside the vm to a single physical interface on the
host to emulate the multiple paths on a single itf laptop for example.

Signed-off-by: Benjamin Hesmans <ben.hesmans@gmail.com>